feat(tools): Adds a beta version of PythonTool and an alpha version of OpenAPI Tool
Adds a beta version of the python tool generator for agent builder.
The python tool can be used to decorate any python function and get a tool which can be used
with the orchestrate agent builder.

Adds an alpha version of the openapi tool generator. This tool allows a user to create
a tool out of an endpoint available from an OpenAPI spec.
Known limitations:
- authentication is not yet implemented
- duplicate parameter names are known to not work
- illegal characters (think body parameters that start with $) are currently not sanitized
- needs to support other content types, such as xml

DCO 1.1 Signoff Eric Marcoux <eric.marcoux@ibm.com>